+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Adds missing thor dependency for CLI
 - CLI automatically sets base-url for http(s) resources
+- Added `--http-headers` CLI option for sending custom headers when fetching the initial SOURCE HTML.
 
 ## [0.0.10] - 2025-08-01